FIX: Clone Node with CSS
Previous implementation of CloneNode relied on using
node.cloneNode(true) to deep copy the nodes and it children.
However this did not copy the css styles from original to cloned.

This commit implements a newer cloneNodeWithCSS function which uses
DFS to clone the DOM node and children applying the styles at each
and every child.